requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Azure SDKs
azure-identity>=1.15.0
azure-cosmos>=4.5.1
azure-servicebus>=7.11.4
azure-keyvault-secrets>=4.7.0
azure-mgmt-resourcegraph>=8.0.0
azure-search-documents>=11.4.0
azure-monitor-opentelemetry>=1.2.0
openai>=1.12.0
gremlinpython>=3.7.0

# Terraform & HCL
python-hcl2>=4.3.2

# Git
GitPython>=3.1.41

# Embeddings & NLP
tiktoken>=0.6.0

# HTTP & Async
httpx>=0.26.0
aiohttp>=3.9.3

# CLI
typer[all]>=0.9.0
rich>=13.7.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.1
PyYAML>=6.0.1
tenacity>=8.2.3
